Tweaks to storeModified to make it comply to the spirit of ds_row
While you can easily add data into modified_data whose keys don't align with any
of the columns in cols, we don't need to push that data back into cols / rows.
DS is meant to represent a row of data and abusing it in that manner doesn't
need to be supported.