Fixing problems when database alias doesn't match the name
When looking up a database from the SQLite tables, DS does look up by the alias.
Unfortunately, it uses the alias name as the database name when storing the
information into dsinfo. This means that if the alias name is not the same as
the database name, queries will produce "1046: No database selected" errors.

This commit fixes this issue by making sure that the database name is pulled
from the "name" field from the "datasource_databases" table.